Refactor API key store module
